Bank Customer Data in VietNam/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
numpy
scikit-learn==1.2.2
tensorflow == 2.11.0